Small code changes

+++ src/config/config_common.c
@@ -33,25 +33,24 @@ void init_config_functions(struct config *config, char *default_lib, char *confi
         log_error_and_exit("%s\n", dlerror());
     }
 
-    if(main_function_name){
+    if(main_function_name) {
         config->main_function = dlsym(config->handle, main_function_name);
         char *error = dlerror();
         if (error != NULL)  {
             log_error_and_exit("\n%s function not found in the provided in library %s. Error from dlsym %s\n", main_function_name, config->library_file_path, error);
         }
-    }
-    else {
+    } else {
         log_error_and_exit("No function name for [%s] provided. Exiting!\n", config_type);
     }
 
-    if(init_function_name){
+    if(init_function_name) {
         config->init_function = dlsym(config->handle, init_function_name);
         if (dlerror() != NULL)  {
             log_error_and_exit("\n%s function not found in the provided in library %s\n", init_function_name, config->library_file_path);
         }
     }
 
-    if(end_function_name){
+    if(end_function_name) {
         config->end_function = dlsym(config->handle, end_function_name);
         if (dlerror() != NULL)  {
             log_error_and_exit("\n%s function not found in the provided in library %s\n", end_function_name, config->library_file_path);
@@ -61,9 +60,9 @@ void init_config_functions(struct config *config, char *default_lib, char *confi
 
 void free_config_data(struct config *cm) {
 
-    ptrdiff_t n = shlen(cm->config_data);
+    size_t n = shlen(cm->config_data);
 
-    for(long i = 0; i < n; i++) {
+    for(size_t i = 0; i < n; i++) {
         free(cm->config_data[i].value);
     }
